Move base api types to pkg/types
