+++ cmd/kubermatic-cluster-controller/cmd/root.go
@@ -22,13 +22,13 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/kubermatic/api"
 	"github.com/kubermatic/api/pkg/controller/cluster"
 	"github.com/kubermatic/api/pkg/controller/version"
 	"github.com/kubermatic/api/pkg/extensions"
 	"github.com/kubermatic/api/pkg/extensions/etcd"
 	"github.com/kubermatic/api/pkg/provider"
 	"github.com/kubermatic/api/pkg/provider/cloud"
+	api "github.com/kubermatic/api/pkg/types"
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
 	"github.com/spf13/viper"
