+++ pkg/controller/cluster/controller.go
@@ -6,7 +6,6 @@ import (
 	"time"
 
 	"github.com/golang/glog"
-	"github.com/kubermatic/api"
 	"github.com/kubermatic/api/pkg/addons/manager"
 	"github.com/kubermatic/api/pkg/controller"
 	"github.com/kubermatic/api/pkg/controller/update"
@@ -15,6 +14,7 @@ import (
 	"github.com/kubermatic/api/pkg/extensions/etcd"
 	"github.com/kubermatic/api/pkg/provider"
 	kprovider "github.com/kubermatic/api/pkg/provider/kubernetes"
+	api "github.com/kubermatic/api/pkg/types"
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
