+++ pkg/controller/version/version_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 
 	"github.com/go-test/deep"
-	"github.com/kubermatic/api"
+	api "github.com/kubermatic/api/pkg/types"
 )
 
 func TestBestAutomaticUpdate(t *testing.T) {
