+++ pkg/handler/datacenter.go
@@ -13,8 +13,8 @@ import (
 	"github.com/go-kit/kit/endpoint"
 	"github.com/golang/glog"
 	"github.com/gorilla/mux"
-	"github.com/kubermatic/api"
 	"github.com/kubermatic/api/pkg/provider"
+	api "github.com/kubermatic/api/pkg/types"
 	"golang.org/x/net/context"
 )
