+++ pkg/provider/cloud/digitalocean/provider.go
@@ -10,10 +10,10 @@ import (
 
 	"github.com/digitalocean/godo"
 	"github.com/golang/glog"
-	"github.com/kubermatic/api"
 	"github.com/kubermatic/api/pkg/extensions"
 	"github.com/kubermatic/api/pkg/provider"
 	ktemplate "github.com/kubermatic/api/pkg/template"
+	api "github.com/kubermatic/api/pkg/types"
 	"github.com/kubermatic/api/pkg/uuid"
 	"golang.org/x/net/context"
 	"golang.org/x/oauth2"
