+++ pkg/provider/cloud/fake/provider.go
@@ -4,9 +4,9 @@ import (
 	"errors"
 	"fmt"
 
-	"github.com/kubermatic/api"
 	"github.com/kubermatic/api/pkg/extensions"
 	"github.com/kubermatic/api/pkg/provider"
+	api "github.com/kubermatic/api/pkg/types"
 	"golang.org/x/net/context"
 	"k8s.io/apimachinery/pkg/util/rand"
 )
