+++ pkg/provider/kubernetes/seed.go
@@ -6,9 +6,9 @@ import (
 	"sync"
 	"time"
 
-	"github.com/kubermatic/api"
 	"github.com/kubermatic/api/pkg/extensions"
 	"github.com/kubermatic/api/pkg/provider"
+	api "github.com/kubermatic/api/pkg/types"
 	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/client-go/pkg/apis/rbac"
 	"k8s.io/client-go/rest"
