+++ pkg/types/certs.go
@@ -1,4 +1,4 @@
-package api
+package types
 
 import (
 	"encoding/base64"
